move fixtures to conftest file
